requirements.txt
streamlit
python-docx
supabase
google-generativeai
